refactor: Duckdb is optional when used as a lib
