+++ src/evaluations/ragas.rs
@@ -4,7 +4,7 @@ use std::path::Path;
 
 use swiftide::query::evaluators::ragas;
 
-use crate::{indexing::build_query_pipeline, repository::Repository, storage};
+use crate::{duckdb::get_duckdb, indexing::build_query_pipeline, repository::Repository};
 
 /// Evaluate a query pipeline with RAGAS
 ///
@@ -60,7 +60,7 @@ pub async fn evaluate_query_pipeline(
     tracing::info!("Building query pipeline");
     let pipeline = build_query_pipeline(
         repository,
-        &storage::get_duckdb(repository),
+        &get_duckdb(repository),
         Some(Box::new(ragas.clone())),
     )?;
