+++ src/lib.rs
@@ -4,7 +4,9 @@ pub mod chat_message;
 pub mod cli;
 pub mod commands;
 pub mod config;
-#[cfg(feature = "evaluations")]
+#[cfg(feature = "duckdb")]
+pub mod duckdb;
+#[cfg(all(feature = "evaluations", feature = "duckdb"))]
 pub mod evaluations;
 pub mod frontend;
 pub mod git;
@@ -12,8 +14,8 @@ pub mod indexing;
 pub mod kwaak_tracing;
 pub mod onboarding;
 pub mod repository;
+#[cfg(feature = "duckdb")]
 pub mod runtime_settings;
-pub mod storage;
 pub mod templates;
 pub mod util;
